Failed scaling implementation. Minor refactor
This commit is left in git history for future reference. All scaling-related changes will be reverted in the next commit.

Failed attempt to implement per-segment scaling, that violates the standard. The standard requires subband weights to be the same for all segments assembled from the same image. This is emphasized by the fact that custom subband weights are specified in segment header part 4, that holds the image/compression specs prohibited to change within one image.

Scaling will be implemented in two parts: subband multiplication by weight factors will be implemented in dwt, and type conversion will be implemented in segment encoding.